content/media/video/test/test_playbackRate.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Media test: playbackRate</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<video id='v1'></video><audio id='a1'></audio>
<pre id="test">
<script class="testbody" type="text/javascript">
var v1 = document.getElementById('v1');
var a1 = document.getElementById('a1');
var passed = true;

is(v1.playbackRate, 1);
is(a1.playbackRate, 1);

v1.playbackRate = 2.5;
a1.playbackRate = 2.5;
is(v1.playbackRate, 2.5);
is(a1.playbackRate, 2.5);

try {
  v1.playbackRate = 0
  a1.playbackRate = 0
  passed = false;
} catch(e) { }
todo(passed, "Should not be able to set playbackRate to 0");

</script>
</pre>
</body>
</html>
